Trying to SOLID
